model: cleanup payment references in PaymentData
Rename PaymentTxnInternalRef to PaymentTransactionExternalKey,
to make it obvious it maps to the external key of the transaction in Kill Bill.

Remove PaymentId and PaymentInternalRef, as these were either not used or misused.

On disk, the plugin stores kb_payment_id, kb_payment_transaction_id and payment_internal_ref,
which now is always the PaymentTransactionExternalKey.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>