Revert "notification: use PSP Reference as transaction external keys"
This reverts commit 396f988a5aa82e6471478c10dc4fda1ac127c424.

The idea of the patch was to make sure we always specify a unique non-blank
external key when creating a transaction (either because of HPP redirect
or chargeback): it is possible we receive '' for example.

Instead of using the PSP reference, we keep using the merchant reference
(if specified during the HPP setup call - for chargebacks it will be most
likely blank), but sanitized. Note that we assume the merchant reference
to be unique (not enforced by the plugin nor Adyen, but will be by Kill Bill
when attempting to create a second transaction for a different HPP for example).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>