Allow passwords (and other configuration entries) to contain '#' character
The plugin expects credential configuration entries in the format
`username#password` (each separated by '|'). This breaks authentication
if username or password contain the character '#'.

While this could be considered uncommon for a username, it could easily
happen for a password, considering that Adyen Customer Area provides
only automatically generated ones.

This commits introduces a simple workaround to mitigate the issue.