Fix billing address validation
As per Adyen documentation, the billing address per se is optional [1],
but when sending it, the country is always mandatory, while the
remaining fields must either be sent all or none [2] (although in our
experience the ZIP always has been treated as optional by Adyen: maybe
it is mandatory only for US and CA).

This commit makes the address validation compliant (altough still only
validates the data presence, not the *values* formal correctness).

[1] https://docs.adyen.com/api-reference/payments-api/paymentrequest/
[2] https://docs.adyen.com/api-reference/common-api/address/